start of refactor to multicart
not sure whether we go down this path of passing a cart handle around (since that IS how template user think of it) or pass around the cart itself.

Leaning towards handle since an incorrect handle will still fetch the first real ordertype. This also simpifies cart service. Would need to make a template var for the hidden field for both cart action and orderTypeHandle.