Round line item salePrice and line item subtotal calculation, just like tax, shipping and discounts on the line item.
This ensures that the total will always be rounded and will infact remove the need for rounding at the order total level. We will remove order level rounding once we add tests.

We will keep the line item’s saleAmount and salePrice unrounded to preserve history and allow purchasable prices to still have more decimal places than the currencies minor units.

This commit fixes a bug where we send the rounded price of a single item (and a qty value) to a payment gatway, but our price total does not match  the gateways calculated total price.