Allow 3rd party order adjusters to insert themselves into the adjustment array.
Usually the adjusters are added to the end of the adjustment process. This change allows them to not just return an array of adjuster classes, but also use an array keyed to insert themselves between the standard adjusters. Eventually we will need to support a UI to manage adjuster classes and ordering.