Include product title in Variant `__toString`
This helps identify variants better, when their element chiclets are displayed in the CMS. Previously, the variant title alone was often confusing (i.e. "Red", rather than "Wrap Dress Red").

I don't think this is a "breaking" change (it still returns a string), but if anyone is relying on the `__toString` method for displaying variant information, they would need to switch to `$variant->title` to get the old string.