Fix for payment processing of free orders
Orders placed through `PaymentsController#actionPay` that do not have an
outstanding balance will still try to send to the payment gateway. This
causes issues with payment gateways that fail due to an invalid amount.
In this instance, commerce-stripe fails to process a payment because the
amount is a non-positive integer.

The root cause of this issue is that the `getIsPaid()` function on the
Order elements is checking if the order is complete to return whether it
was paid. I was afraid to remove the check of `isComplete` out of that
function because it seemed to be there intentionally.

To fix this, I added a new public function to the order element
`hasOutstandingBalance()` this actually helps remove some duplicated
conditionals throughout the codebase. Finally, I updated the payments
service to utilize this method instead of the `getIsPaid`.