Changed an incorrectly used <label> to a <legend> with a new <fieldset> element.
Added a few missing <label> elements. (The use of a placeholder attribute does not negate the need for a <label> element.)

Added a <span> to wrap the text content of more <label>s and added a label class to such new and existing <span> elements.

Added an <abbr> element for the "Qty" table columns.

Added an appropriate scope attribute to <th> elements.

Replaced most <br> elements with one of two classes ("block" and "mb-line").

Corrected a typo in an existing aria-label attribute.