Optionally allow a shippingMethod to be associated with a ShippingMethodOption
Resolves #3271

Craft Commerce's ShippingMethodOption erases the information of any
custom ShippingMethod's that were used to create them. While not ideal,
this change sets a new field, shippingMethod on the ShippingMethodOption
class and also proxies the call to getShippingRules to use the attached
ShippingMethod if it is present.

Of course, other functions such as getName, getType, etc will not be
overridden and would need to be accessed like
$option->shippingMethod->getType().